requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
requests>=2.32.0
python-multipart>=0.0.20
openai>=1.51.0
python-dotenv>=1.0.0
llama-cpp-python>=0.3.0
